Refactor: Kakao관련 useState, useEffect훅 커스텀훅으로 관리및 타입에러 해결
